+++ marginaleffects/model_sklearn.py
@@ -11,28 +11,16 @@ class ModelSklearn(ModelAbstract):
     def __init__(self, model, vault={}):
         super().__init__(model, vault)
 
-    def get_predict(self, params, newdata: pl.DataFrame):
+    def get_predict(self, params, newdata):
         engine = self.get_engine_running()
+        formula = self.get_formula()
 
         if isinstance(newdata, np.ndarray):
             exog = newdata
+        elif callable(formula):
+            _, exog = formula(newdata)
         else:
-            try:
-                import formulaic
-
-                if isinstance(newdata, formulaic.ModelMatrix):
-                    exog = newdata.to_numpy()
-                else:
-                    if isinstance(newdata, pl.DataFrame):
-                        nd = newdata.to_pandas()
-                    else:
-                        nd = newdata
-                    y, exog = formulaic.model_matrix(self.model.formula, nd)
-                    exog = exog.to_numpy()
-            except ImportError:
-                raise ImportError(
-                    "The formulaic package is required to use this feature."
-                )
+            _, exog = model_matrices(formula, data=newdata)
 
         try:
             with warnings.catch_warnings():
@@ -73,9 +61,7 @@ class ModelSklearn(ModelAbstract):
 
 
 # @validate_types
-def fit_sklearn(
-    formula, data: pl.DataFrame, engine, kwargs_engine={}, kwargs_fit={}
-) -> ModelSklearn:
+def fit_sklearn(formula, data: pl.DataFrame, engine) -> ModelSklearn:
     """
     Fit a sklearn model with output that is compatible with pymarginaleffects.
 
@@ -97,7 +83,10 @@ def fit_sklearn(
     elif callable(formula):
         y, X = formula(d)
 
-    engine_running = engine(**kwargs_engine).fit(X=X, y=y, **kwargs_fit)
+    else:
+        raise ValueError("The formula must be a string or a callable function.")
+
+    engine_running = engine.fit(X=X, y=y)
 
     vault = {
         "formula": formula,
@@ -131,25 +120,62 @@ This function streamlines the process of fitting sklearn models by:
 """
     + DocsModels.docstring_kwargs_engine
     + """
-`kwargs_fit` : (dict, default={}) Additional arguments passed to the model's fit method. 
 """
     + DocsModels.docstring_fit_returns("Sklearn")
     + """
 ## Examples
 
-```python
-from sklearn.linear_model import LinearRegression
+```{python}
 from marginaleffects import *
+from statsmodels.formula.api import ols
+import polars.selectors as cs
+from sklearn.pipeline import make_pipeline
+from sklearn.model_selection import train_test_split
+from sklearn.preprocessing import OneHotEncoder, FunctionTransformer
+from sklearn.linear_model import LinearRegression
+from sklearn.compose import make_column_transformer
+from xgboost import XGBRegressor
 
-data = get_dataset("thornton")
 
-model = fit_sklearn(
-    formula="outcome ~ distance + incentive",
-    data=data,
-    engine=LinearRegression,
+# Linear regression: Scikit-learn
+military = get_dataset("military")
+
+mod_sk = fit_sklearn(
+    "rank ~ officer + hisp + branch",
+    data=military,
+    engine=LinearRegression(),
 )
+avg_predictions(mod_sk, by="branch")
 
-predictions(model)
+# Linear regression: Statsmodels
+mod_sm = ols("rank ~ officer + hisp + branch", data=military.to_pandas()).fit()
+avg_predictions(mod_sm, by="branch")
+
+# XGBoost: Scikit-learn
+airbnb = get_dataset("airbnb")
+
+train, test = train_test_split(airbnb)
+
+catvar = airbnb.select(~cs.numeric()).columns
+
+
+def selector(data):
+    y = data.select(cs.by_name("price", require_all=False))
+    X = data.select(~cs.by_name("price", require_all=False))
+    return y, X
+
+
+preprocessor = make_column_transformer(
+    (OneHotEncoder(), catvar),
+    remainder=FunctionTransformer(lambda x: x.to_numpy()),
+)
+pipeline = make_pipeline(preprocessor, XGBRegressor())
+
+mod = fit_sklearn(selector, data=train, engine=pipeline)
+
+avg_predictions(mod, newdata=test, by="unit_type")
+
+avg_comparisons(mod, variables={"bedrooms": 2}, newdata=test)
 ```
 """
     + DocsModels.docstring_notes("sklearn")