+++ tests/test_datagrid.py
@@ -104,3 +104,13 @@ def test_callable():
 
     p = predictions(mod, newdata=datagrid(mpg=fivenum))
     assert p.height == 5
+
+
+def test_datagrid_counterfactual_lambda():
+    mtcars = get_dataset("mtcars", "datasets")
+    grid = datagrid(
+        am = lambda x: x.unique(),
+        cyl = np.unique,
+        newdata = mtcars,
+        grid_type = "counterfactual")
+    assert grid.height == 32 * 2 * 3